refactor: Enum 구조 변경
- Lotto 설정 관련 상수들을 LottoConfig로 통합하였음
- Rank 관련 상수들을 WiningRank로 통합하였음
