--- src/main/java/lotto/config/LottoPrice.java
@@ -1,15 +1,0 @@
-package lotto.config;
-
-public enum LottoPrice {
-    LOTTO_PRICE(1000);
-
-    private final int price;
-
-    LottoPrice(int price) {
-        this.price = price;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-}
